Add sys_admins SSH keys to Hetzner project
This grants root SSH access to the servers created in the Hetzner
account which allows them to perform all sorts of admin management
tasks.

This requires specifying a Hetzner API token for the Ansible role to
work. You'll see that token named "Ansible" in the API tokens section of
the Cloud Console account.