Invalidate existing circleci pip packages cache
We changed the directory path that gets cached but we didn't invalidate
it which makes us think this is the reasone why `pip` is not picking it
up and re-downloading the dependencies.

Checking with @danypr92 it seems true that dependencies end up in
`/usr/local/lib/python3.7/site-packages` so the `save_cache` `paths`
parameter should be good.