Serve app traffic from HTTP
This is a follow up from https://github.com/coopdevs/donalo/pull/23 not
only we need to disable HTTPS (until we fix it) but we also need to
serve app traffic meanwhile. Donalo wants to play with the app.

Note that we also replace the socket with a binding to port 8080,
Unicorn's default. This is because so far I only managed to boot it
manually running `bundle exec unicorn`.

This depends on https://github.com/coopdevs/sharetribe/pull/6.